requirements.txt
beautifulsoup4==4.4.1
cssselect==0.9.1
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.0
jieba3k==0.35.1
lxml==3.5.0
newspaper3k==0.1.7
nltk==3.1
Pillow==3.1.1
py==1.4.31
pytest==2.8.7
python-dateutil==2.4.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
tldextract==1.7.5
vcrpy==1.7.4
wrapt==1.10.6
